separate server connection from app
